+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/command/api/GetPaymentDetailCommand.java
@@ -1,9 +1,10 @@
-package com.taken_seat.payment_service.application.command;
+package com.taken_seat.payment_service.application.command.api;
 
 import java.util.UUID;
 
+import com.taken_seat.payment_service.application.command.PaymentCommand;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
-import com.taken_seat.payment_service.application.service.PaymentService;
+import com.taken_seat.payment_service.application.service.api.PaymentService;
 
 public class GetPaymentDetailCommand implements PaymentCommand<PaymentDetailResDto> {
