+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/command/api/RegisterPaymentCommand.java
@@ -1,8 +1,9 @@
-package com.taken_seat.payment_service.application.command;
+package com.taken_seat.payment_service.application.command.api;
 
+import com.taken_seat.payment_service.application.command.PaymentCommand;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentDto;
-import com.taken_seat.payment_service.application.service.PaymentService;
+import com.taken_seat.payment_service.application.service.api.PaymentService;
 
 public class RegisterPaymentCommand implements PaymentCommand<PaymentDetailResDto> {
