+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/presentation/controller/PaymentController.java
@@ -18,18 +18,18 @@ import com.taken_seat.common_service.aop.annotation.RoleCheck;
 import com.taken_seat.common_service.aop.vo.Role;
 import com.taken_seat.common_service.dto.ApiResponseData;
 import com.taken_seat.common_service.dto.AuthenticatedUser;
-import com.taken_seat.payment_service.application.command.DeletePaymentCommand;
-import com.taken_seat.payment_service.application.command.GetPaymentDetailCommand;
-import com.taken_seat.payment_service.application.command.RegisterPaymentCommand;
-import com.taken_seat.payment_service.application.command.SearchPaymentCommand;
-import com.taken_seat.payment_service.application.command.UpdatePaymentCommand;
+import com.taken_seat.payment_service.application.command.api.DeletePaymentCommand;
+import com.taken_seat.payment_service.application.command.api.GetPaymentDetailCommand;
+import com.taken_seat.payment_service.application.command.api.RegisterPaymentCommand;
+import com.taken_seat.payment_service.application.command.api.SearchPaymentCommand;
+import com.taken_seat.payment_service.application.command.api.UpdatePaymentCommand;
 import com.taken_seat.payment_service.application.dto.controller.request.PaymentRegisterReqDto;
 import com.taken_seat.payment_service.application.dto.controller.request.PaymentUpdateReqDto;
 import com.taken_seat.payment_service.application.dto.controller.response.PagePaymentResponseDto;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentSearchDto;
-import com.taken_seat.payment_service.application.service.PaymentService;
+import com.taken_seat.payment_service.application.service.api.PaymentService;
 import com.taken_seat.payment_service.infrastructure.mapper.PaymentMapper;
 import com.taken_seat.payment_service.infrastructure.swagger.PaymentSwaggerDocs;
